chore: organize generated data files
